lab 2/requirements.txt
# Dependências do Projeto - Lab 02: Análise de Qualidade de Repositórios Java

# Manipulação de dados e análise
pandas>=1.5.0
numpy>=1.21.0

# Requisições HTTP para API do GitHub
requests>=2.28.0

# Configuração e variáveis de ambiente
python-dotenv>=0.19.0

# Para futuras análises estatísticas e visualizações (Sprint 2)
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.8.0

# Utilitários adicionais
pathlib2>=2.3.0; python_version < '3.4'
